+++ src/MicroElements.OpenApi.FluentValidation/AssemblyAttributes.cs
@@ -2,4 +2,5 @@
 
 [assembly: InternalsVisibleTo("MicroElements.Swashbuckle.FluentValidation")]
 [assembly: InternalsVisibleTo("MicroElements.NSwag.FluentValidation")]
+[assembly: InternalsVisibleTo("MicroElements.AspNetCore.OpenApi.FluentValidation")]
 [assembly: InternalsVisibleTo("MicroElements.Swashbuckle.FluentValidation.Tests")]